requirements.txt
flask==2.3.2
tensorflow==2.13.0
numpy==1.24.4
nibabel==5.1.0
scipy==1.11.2
Pillow==10.0.0